update to recent version

+++ app/code/Magento/Customer/Model/ResourceModel/Customer/Relation.php
@@ -23,41 +23,43 @@ class Relation implements \Magento\Framework\Model\ResourceModel\Db\VersionContr
         $defaultBillingId = $customer->getData('default_billing');
         $defaultShippingId = $customer->getData('default_shipping');
 
-        /** @var \Magento\Customer\Model\Address $address */
-        foreach ($customer->getAddresses() as $address) {
-            if ($address->getData('_deleted')) {
-                if ($address->getId() == $defaultBillingId) {
-                    $customer->setData('default_billing', null);
-                }
+        if (!$customer->getData('ignore_validation_flag')) {
+            /** @var \Magento\Customer\Model\Address $address */
+            foreach ($customer->getAddresses() as $address) {
+                if ($address->getData('_deleted')) {
+                    if ($address->getId() == $defaultBillingId) {
+                        $customer->setData('default_billing', null);
+                    }
 
-                if ($address->getId() == $defaultShippingId) {
-                    $customer->setData('default_shipping', null);
-                }
+                    if ($address->getId() == $defaultShippingId) {
+                        $customer->setData('default_shipping', null);
+                    }
 
-                $removedAddressId = $address->getId();
-                $address->delete();
+                    $removedAddressId = $address->getId();
+                    $address->delete();
 
-                // Remove deleted address from customer address collection
-                $customer->getAddressesCollection()->removeItemByKey($removedAddressId);
-            } else {
-                $address->setParentId(
-                    $customer->getId()
-                )->setStoreId(
-                    $customer->getStoreId()
-                )->setIsCustomerSaveTransaction(
-                    true
-                )->save();
+                    // Remove deleted address from customer address collection
+                    $customer->getAddressesCollection()->removeItemByKey($removedAddressId);
+                } else {
+                    $address->setParentId(
+                        $customer->getId()
+                    )->setStoreId(
+                        $customer->getStoreId()
+                    )->setIsCustomerSaveTransaction(
+                        true
+                    )->save();
 
-                if (($address->getIsPrimaryBilling() ||
-                        $address->getIsDefaultBilling()) && $address->getId() != $defaultBillingId
-                ) {
-                    $customer->setData('default_billing', $address->getId());
-                }
+                    if (($address->getIsPrimaryBilling() ||
+                            $address->getIsDefaultBilling()) && $address->getId() != $defaultBillingId
+                    ) {
+                        $customer->setData('default_billing', $address->getId());
+                    }
 
-                if (($address->getIsPrimaryShipping() ||
-                        $address->getIsDefaultShipping()) && $address->getId() != $defaultShippingId
-                ) {
-                    $customer->setData('default_shipping', $address->getId());
+                    if (($address->getIsPrimaryShipping() ||
+                            $address->getIsDefaultShipping()) && $address->getId() != $defaultShippingId
+                    ) {
+                        $customer->setData('default_shipping', $address->getId());
+                    }
                 }
             }
         }
